WIP: redirect to blog with a new parameter
